Revert "change to using CdiDecorator approach"
This reverts commit e82f570de53f0e9eee0c194552d54c7eb76f37d4.
